Implement object Rack, which is a hardware object for the "stepper motor + 2 limit switches". (Made a small driving test with both stepper motor and limit switches accounted for)
Consider the essential elements of a Rack's operation to be: the pins to read from the limit switches,
the STEP and DIR pins, which are specific to driving the stepper motor through the driver (A4988).

Observed that the state of pin DIR (true/false) determines the direction in which the driver
'intends' to drive the motor. Therefore, it seems both natural and convenient for a Rack object to keep a
note of the state of the pin DIR.

Observation:
A program can drive the motor 1 step by giving a definitive low-to-high pulse. "Definitive" here means
that the low-to-high transition must be discernible. It has been examined to show that a 1ms low state
followed by a to-high transition, without the high state persisting long enough after that, fails to
drive the motor.
Implement:
Chose the signal pattern of [ 1ms off - 1ms on ] to drive the motor 1 step.
Observation:
Different motor (different Rack objects) might have different time pattern for the driving signal.
Conclusion:
Rack object keeps a time variable (type TickType_t). The variable is mean to indicate the amount of time
each off and on signal should persist, such as 1ms in [ 1ms off - 1ms on].